Add queued message notifier channel
