+++ lcm/src/lcm/receiver.rs
@@ -1,46 +1,29 @@
-use std::io;
-use std::net::{UdpSocket, SocketAddr, IpAddr, Ipv4Addr};
+use std::sync::mpsc;
+use std::net::UdpSocket;
 
 /// The LCM backend used for receiving messages without blocking the main
 /// thread.
 pub struct Receiver {
+    /// The multicast socket used for receiving datagrams.
     socket: UdpSocket,
+
+    /// The channel used to notify the `Lcm` object that messages have been
+    /// queued.
+    notify_tx: mpsc::SyncSender<()>,
 }
 impl Receiver {
     /// Create a `Receiver` with the specified settings.
-    pub fn new(addr: &Ipv4Addr, port: u16, ttl: u32) -> io::Result<Self> {
-        let socket = Receiver::setup_udp_socket(addr, port, ttl)?;
-
-        Ok(Receiver {
-            socket,
-        })
+    pub fn new(socket: UdpSocket, notify_tx: mpsc::SyncSender<()>) -> Self {
+        Receiver { socket, notify_tx }
     }
 
-    /// Set up the UDP socket.
-    fn setup_udp_socket(addr: &Ipv4Addr, port: u16, ttl: u32) -> io::Result<UdpSocket> {
-        debug!("Binding UDP socket");
-        let socket = {
-            let inaddr_any = IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0));
-            UdpSocket::bind(&SocketAddr::new(inaddr_any, port))?
-        };
-
-        // FIXME:
-        // Rust only supportes SO_REUSEADDR through crates. Come back to this
-        // later, after deciding which crate to use (probably net2 or plain
-        // libc). See lcm_udpm.c:936-956
-        warn!("Skipping SO_REUSEADDR and SO_REUSEPORT");
-
-        // FIXME:
-        // The C version of LCM increases the receive buffer size on Win32. Do
-        // we need to do this and how?
-        warn!("Not checking receive buffer size");
-
-        debug!("Joining multicast group");
-        socket.join_multicast_v4(addr, &Ipv4Addr::new(0, 0, 0, 0))?;
-
-        debug!("Setting multicast packet TTL to {}", ttl);
-        socket.set_multicast_ttl_v4(ttl)?;
-
-        Ok(socket)
+    /// Enter the `Receiver` execution loop.
+    ///
+    /// This function will wait for events on the UDP socket and forward them
+    /// through the appropriate channels based on subscriptions. It will only
+    /// exit if the subscription channel closes (which signifies that the
+    /// client `Lcm` object has been deleted).
+    pub fn run(&self) {
+        
     }
 }
